dashboard/requirements.txt
streamlit
pandas
pyodbc
azure-identity
azure-keyvault-secrets
sqlalchemy